### 2026-03-20 #10236 @editorialbot
[PRE REVIEW]: SARXarray: Xarray extension for Synthetic Aperture Radar data
**Submitting author:** @rogerkuou (<a href="http://orcid.org/0000-0002-5373-5209">Ou Ku</a>)
**Repository:** https://github.com/TUDelftGeodesy/sarxarray
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.3
