### 2026-03-25 #10236 @rwegener2
👋🏻 @jiangyuinsar and @bcankara, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
