### 2026-03-26 #10236 @bcankara
@rwegener2 Thank you for considering me as a potential reviewer for this submission. SARXarray is closely related to my research area, and I appreciate being suggested. Unfortunately, due to my current workload and commitments, I am unable to take on this review at this time. I hope you find a suitable reviewer soon, and I wish the authors success with their work.
