### 2026-04-21 952f30d
Add cropping function
