### 2026-04-22 696418f
Add crop test, allow for non-integer crops
