### 2026-04-22 06afd01
Remove da_crop reference in ValueError test
