### 2026-04-22 0146f22
Add option to directly pass the bounding box as a tuple
