### 2026-04-25 #10236 @gmarupilla
@rwegener2

I would like to volunteer to review this submission.

Relevant background:
- Areas of overlap: xarray, geospatial / remote sensing data pipelines
- Recent peer-review activity: JOSS, PLOS Computational Biology, IEEE S&P AEC
- Domain experience: geospatial and remote sensing data, FAIR + reproducibility
