### 2026-04-27 #10236 @rwegener2
@gmarupilla thanks for volunteering! Do you have any experience working specifically with SAR data?
