### 2026-04-28 #10236 @gmarupilla
@rwegener2 Thanks for asking. I do not have hands-on project experience specifically with SAR data yet.

My relevant experience is adjacent: I have worked with geospatial/remote-sensing data pipelines, reproducible scientific software, xarray-style data workflows, and FAIR/reproducibility review. I would be comfortable reviewing packaging, documentation, testing, API usability, reproducibility, and general geospatial data-handling aspects of the submission.

If the review requires deep SAR-domain validation, I may be better as a secondary reviewer alongside someone with direct SAR/InSAR expertise.
