### 2026-05-01 #10236 @rwegener2
Thanks so much for volunteering @gmarupilla. Being familiar with xarray workflows is very helpful. Luckily @weiji14 has experience with SAR.
