### 2026-05-07 TUDelftGeodesy/sarxarray#87 @gmarupilla
Document optional extras ([demo], [dev], [docs]) in README install section
The README install section currently shows only:

```sh
pip install sarxarray
```

`pyproject.toml` defines three useful extras (`[dev]`, `[docs]`, `[demo]`) but they aren't mentioned anywhere in the README. Users following the tutorial notebook hit `ModuleNotFoundError: matplotlib` at the visualisation cells because `jupyterlab` and `matplotlib` only come in via `[demo]`.

Two small README additions would help:

1. List the available extras and what each one is for.
2. Show the zsh-safe form, since `pip install sarxarray[demo]` errors out as `zsh: no matches found: sarxarray[demo]` on the default macOS shell:

   ```sh
   pip install 'sarxarray[demo]'    # to run the example notebook
   pip install 'sarxarray[dev]'     # to run the test suite
   pip install 'sarxarray[docs]'    # to build the docs locally
   ```

Happy to send a PR if useful.

(Filed in the context of JOSS review openjournals/joss-reviews#10492.)
