### 2026-05-07 TUDelftGeodesy/sarxarray#88 @gmarupilla
Cross-ref: this is one of the failure modes triggered by #89 (the tutorial notebook expects a hand-downloaded dataset, and an empty file list lands here).
