### 2026-05-07 #10492 @gmarupilla
Hi @rogerkuou — really nice submission. The test suite (48 tests, 89% coverage) passes cleanly, the paper is well structured, and SARXarray's place in the Pangeo / Xarray ecosystem is well argued.

A few things came up while running through the tutorial that I think are worth fixing, but none of them are checklist failures — I've filed them as issues on the upstream repo so they live in the right place:

- TUDelftGeodesy/sarxarray#87 — README doesn't mention the `[demo]` / `[dev]` / `[docs]` extras. New users on macOS zsh also hit the bracket-globbing gotcha (`zsh: no matches found: sarxarray[demo]`).
- TUDelftGeodesy/sarxarray#89 — the tutorial notebook expects the figshare dataset to be downloaded and unzipped by hand. A small `urllib` / `zipfile` cell (or `pooch`) would make it one-click reproducible. Once the dataset is in place the notebook runs end-to-end on Python 3.14.4 / xarray 2026.4 / dask 2026.3.
- TUDelftGeodesy/sarxarray#88 — `from_binary([])` produces a confusing xarray-internal error instead of a clear "no SLC files provided" message. An early-guard one-liner would do it.

Two smaller things to note inline rather than as issues:

- `paper.bib`: the `ozer2018applicability` entry is missing volume/issue (currently just `pages = {e12509}`).
- Coverage on `sarxarray/_io.py` is 85% — lower than the 92–100% on the other modules. The I/O surface is the most user-facing bit, so a couple more tests around edge cases of the binary reader would be nice to have, though not required.

Happy to chat through any of the above. Otherwise this is good to go from my side.
