### 2026-05-18 89a7a8f
update README with extra dependencies and zsh safe format
