+++ README.md
@@ -16,7 +16,7 @@ SARXarray is an open-source Xarray extension for Synthetic Aperture Radar (SAR) 
 SARXarray can be installed from PyPI:
 
 ```sh
-pip install sarxarray
+pip install 'sarxarray'
 ```
 
 or from the source:
@@ -28,6 +28,16 @@ pip install .
 ```
 
 Note that Python version `>=3.10` is required for SARXarray.
+
+### Install extra dependencies
+
+SARXarray has some optional dependencies to support additional functionality besides the core features. You can install them depending on your needs:
+
+```sh
+pip install 'sarxarray[demo]'    # to run the demo notebook
+pip install 'sarxarray[dev]'     # to run the test suite
+pip install 'sarxarray[docs]'    # to build the docs locally
+```
 
 ## Documentation
 
