### 2026-05-18 1d32fff
update installation guide in docs
