### 2026-05-18 36f3505
handle empty slc list and str slc
