### 2026-05-18 3c57792
annotation for from_binary
