### 2026-05-18 2b120d5
remove redundant driver check
