### 2026-05-18 0e2a24e
remove redundant driver check
