### 2026-05-18 4035a30
remove trying to read meta array keys from ifg.res
