### 2026-05-18 8e65750
remove list ensurance since this is already done in read_metadata
