### 2026-05-18 d4d7d24
test more edge cases in io
