### 2026-05-18 7d4e8da
Apply suggestions from code review
Co-authored-by: Simon-van-Diepen <44979564+Simon-van-Diepen@users.noreply.github.com>
