### 2026-05-18 c6cea98
Update liscense text in README
