### 2026-05-18 TUDelftGeodesy/sarxarray#88 @rogerkuou
Hi @gmarupilla, PR #91 address both this issue, as well as your comment on low test coverage. We added some new tests on the edge cases (mainly on metadata reading). The new coverage is now 95%.

Again, thanks for raising this issue and feel free to comment!
