### 2026-05-18 TUDelftGeodesy/sarxarray#89 @rogerkuou
Hi @gmarupilla , thanks for already giving a solution!

I added the downloading code as a commented out cell in the demo notebook. Eventually I did not go with pooch for a simplier dependency. The updated notebook is already readered in the doc website: https://tudelftgeodesy.github.io/sarxarray/notebooks/demo_sarxarray/ .

With this I will close this issue for now. Feel free to leave further comments!
