### 2026-05-19 TUDelftGeodesy/sarxarray#88 @gmarupilla
Thank you, LGTM, let me know when it merged, I can test it, but overall MR looks good to me
