### 2026-05-19 35fd1b4
fix doi links in joss paper
