+++ paper/paper.bib
@@ -72,7 +72,7 @@ volume = {12},
 number = {S2},
 pages = {e12509},
 keywords = {InSAR, levee deformation, levee monitoring, levee safety, satellite radar interferometry},
-doi = {https://doi.org/10.1111/jfr3.12509},
+doi = {10.1111/jfr3.12509},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/jfr3.12509},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/jfr3.12509},
 year = {2019}
@@ -96,7 +96,7 @@ volume = {463},
 pages = {117551},
 year = {2025},
 issn = {0016-7061},
-doi = {https://doi.org/10.1016/j.geoderma.2025.117551},
+doi = {10.1016/j.geoderma.2025.117551},
 url = {https://www.sciencedirect.com/science/article/pii/S0016706125003921},
 author = {Yustisi Lumban-Gaol and Philip Conroy and Simon {van Diepen} and Freek {van Leijen} and Ramon Hanssen},
 keywords = {InSAR, Pastures, Peatlands, SPAMS, Subsidence},
