### 2026-05-20 TUDelftGeodesy/sarxarray#88 @gmarupilla
> Hi [@gmarupilla](https://github.com/gmarupilla), FYI PR [#91](https://github.com/TUDelftGeodesy/sarxarray/pull/91) has been merged.

I'll take a look in few hours, thank you
