### 2026-05-20 TUDelftGeodesy/sarxarray#89 @gmarupilla
Thanks @rogerkuou — the commented-out cell is a clean solution and keeps the dependency surface small. Works for me too, closing from my side.
