### 2026-05-20 TUDelftGeodesy/sarxarray#88 @gmarupilla
Pulled latest `main` and checked PR #91 — the early guard on empty/non-iterable `slc_files` raises a clear `ValueError` now instead of falling into the xarray internals. The extra tests around the metadata reader are a nice bonus, and the coverage bump on `_io.py` is great to see. Thanks for the quick turnaround!
