### 2026-05-23 TUDelftGeodesy/sarxarray#94 @weiji14
Document release process and include actual release notes in Changelog file
The release process doesn't appear to be documented very well (e.g. with a checklist). The past few release PRs did not have without any description (e.g. #68, #73, #76), and the changelog page at https://tudelftgeodesy.github.io/sarxarray/CHANGELOG/ simply points to https://github.com/TUDelftGeodesy/sarxarray/releases even though it states "All notable changes to this project will be documented in this file".

https://github.com/TUDelftGeodesy/sarxarray/blob/3a11cd07450045b74a8418135d2aa949f32a87fa/docs/CHANGELOG.md?plain=1#L1-L6

Recommendations:
1. Actually include release notes in CHANGELOG.md (refer to https://www.pyopensci.org/python-package-guide/documentation/repository-files/changelog-file.html on formatting)
2. Create a checklist, possibly as an Issue/Pull Request template, so that maintainers know how to create a release. Include steps such as:
    - [ ] Update CHANGELOG.md file
    - [ ] Update version in `pyproject.toml`
    - [ ] Create release/tag at https://github.com/TUDelftGeodesy/sarxarray/releases

Note that these are only suggestions, and you're welcome to implement some or all of the above 🙂 There are also other 'advanced' methods of e.g. updating the version based on git tags (https://www.pyopensci.org/python-package-guide/package-structure-code/python-package-versions.html#tool-3-setuptools-scm-versioning-using-git-tags), but up to you on what you're comfortable with implementing.

Part of JOSS review at https://github.com/openjournals/joss-reviews/issues/10492
