### 2026-05-23 TUDelftGeodesy/sarxarray#95 @weiji14
Missing API documentation for some public functions
Spotted some public functions (those not starting with an underscore) that don't appear to be documented yet at https://tudelftgeodesy.github.io/sarxarray/api_reference/:

- [I/O module](https://github.com/TUDelftGeodesy/sarxarray/blob/3dcde0e8a9d016e5c939f4bc5ac3efea90c4a0fa/docs/api_reference.md?plain=1#L13-L17):
  - [`read_metadata` function](https://github.com/TUDelftGeodesy/sarxarray/blob/v1.2.3/sarxarray/_io.py#L288-L373). Xref https://github.com/TUDelftGeodesy/sarxarray/issues/82
- [Utility](https://github.com/TUDelftGeodesy/sarxarray/blob/v1.2.3/docs/api_reference.md?plain=1#L19-L23)
  - [`crop` function](https://github.com/TUDelftGeodesy/sarxarray/blob/main/sarxarray/utils.py#L158-L208), added in https://github.com/TUDelftGeodesy/sarxarray/pull/86

Could either manually add entries for these to the `docs/api_reference.md` file, or to avoid situations like this in the future, use something like [autodoc](https://mkdocstrings.github.io/usage/#autodoc-syntax) that will ensure public functions in a module are always documented.

Side note: May also want to update these lines in the `__init__.py` file?

https://github.com/TUDelftGeodesy/sarxarray/blob/3a11cd07450045b74a8418135d2aa949f32a87fa/sarxarray/__init__.py#L1-L3

Not sure which functions you want to allow usage from top-level (e.g. a user calls `sarxarray.crop` instead of `sarxarray.utils.crop`), will leave it up to you to decide on the namespace.

*Part of JOSS review at https://github.com/openjournals/joss-reviews/issues/10492*
