### 2026-05-26 8bfd828
Add binary file writer
