### 2026-05-26 bfaff9a
Fix docstring
