### 2026-05-27 b19ca78
Fix linting, add tests
