### 2026-05-27 403e910
Split filewriting test in dataset and dataarray
