### 2026-05-28 0cdc6ed
Add input arg to control the permissions of overwriting files
