### 2026-06-02 TUDelftGeodesy/sarxarray#99 @weiji14
Minor clarification on sarxarray accessor vs utillity methods
Just want to clarify something from this [line](https://github.com/TUDelftGeodesy/sarxarray/blob/a08d7d67f3304df3d568afacb9d0f6042e74fd34/paper/paper.md?plain=1#L53) in the paper draft

> The software has three main components: an I/O module that lazily loads binary SLCs and related metadata, a Stack accessor that provides SAR-specific operations (amplitude/phase extraction, multi-looking, Mean-Reflection-Map generation, coherence calculation), and utility functions.

From what I can tell, the `Stack` accessor only provides these methods:
- [`mrm`](https://tudelftgeodesy.github.io/sarxarray/api_reference/#sarxarray.stack.Stack.mrm) (Mean-Reflection-Map ✅)
- [`multi_look`](https://tudelftgeodesy.github.io/sarxarray/api_reference/#sarxarray.stack.Stack.multi_look) (multi-looking ✅)
- [`point_selection`](https://tudelftgeodesy.github.io/sarxarray/api_reference/#sarxarray.stack.Stack.point_selection) (not mentioned ❓)

The amplitude/phase extraction appears to be inherent to the xarray data structure (data loaded as a numpy complex64 dtype). Coherence calculation appears to be under [utils.complex_coherence](https://tudelftgeodesy.github.io/sarxarray/api_reference/#sarxarray.utils.complex_coherence). Could the sentence be rephrased perhaps to be clear on what is an accessor method (accessed through `.slcstack.<something>`) vs a util function (e.g. `sarxarray.utils.complex_coherence(...)`).

Part of https://github.com/openjournals/joss-reviews/issues/10492
