### 2026-06-02 #10492 @weiji14
Ok, have finished a second round of reviews, focusing mostly on the 'Functionality' and 'Software Design' aspect. Have opened one more [issue](https://github.com/TUDelftGeodesy/sarxarray/issues/99) to clarify a minor point on this.

One more thing I'll need to check through are the references. Will do that once there is an updated PDF generated.
