### 2026-06-05 807b77c
update mamba installation link
