### 2026-06-05 6a69498
add read_metadata to_bianry and crop to api doc
