+++ docs/api_reference.md
@@ -16,8 +16,14 @@ Here is the API reference for the `sarxarray` package.
 
 ::: sarxarray._io.from_binary
 
+::: sarxarray._io.read_metadata
+
+:: sarxarray._io.to_binary
+
 ## **Utility**
 
 ::: sarxarray.utils.multi_look
 
 ::: sarxarray.utils.complex_coherence
+
+::: sarxarray.utils.crop
