### 2026-06-05 TUDelftGeodesy/sarxarray#95 @rogerkuou
Hi @weiji14 , thanks for pointing this out!
The API documentation for three more functions is added by #102 . During the time of reviewing, @Simon-van-Diepen already updated the `__init__.py` file:

https://github.com/TUDelftGeodesy/sarxarray/blob/8bfd828a611f92b23ab0147a7722692aa4b6a734/sarxarray/__init__.py#L1-L3

After fixing all current Joss reviewing issues, I will make a new release and point the JOSS submission to the new release.

Also thanks for suggesting `autodoc`! Can be super useful! For now I documented it in #101 , since we are recently considering switching to ReadthDoc + Sphinx, due to the shadow future of `MKDocs`...
