### 2026-06-24 7825af5
Update docs/api_reference.md
Co-authored-by: Simon-van-Diepen <44979564+Simon-van-Diepen@users.noreply.github.com>
