+++ docs/api_reference.md
@@ -18,7 +18,7 @@ Here is the API reference for the `sarxarray` package.
 
 ::: sarxarray._io.read_metadata
 
-:: sarxarray._io.to_binary
+::: sarxarray._io.to_binary
 
 ## **Utility**
 
