### 2026-06-24 59e9050
add agent skills for generating CHANGELOG
