### 2026-06-24 397a6e8
add a issue template
