### 2026-06-24 b10b7e0
update ai usge disclosure section in JOSS paper
