### 2026-06-24 TUDelftGeodesy/sarxarray#94 @rogerkuou
Hi @weiji14 , sorry for being a bit silent recently, was occupied by other tasks.

I just worked on this issue via #103 . While addressing your awesome suggestions, I could not help to think if LLM agent can help with generating CHANGELOG info. I implemented an agent skill for this, but still leave reviewing and committing to human.

While my colleague is reviewing this PR, feel free to share your thoughts!
