### 2026-06-24 c2e0fcf
add deprecation warning for point_selection and update unite tests
