### 2026-06-24 TUDelftGeodesy/sarxarray#99 @rogerkuou
Hi @weiji14 , thanks for the nice suggestion! I have rephrased this sentence via PR #104 , it should reflect the reality now.

Regarding [point_selection](https://tudelftgeodesy.github.io/sarxarray/api_reference/#sarxarray.stack.Stack.point_selection), this is a legacy function introduced in the early stage of development. It selects the Persistent Scatterers, and [has been moved to DePSI](https://github.com/TUDelftGeodesy/DePSI/blob/944d664bb44321e4c1f6455518cda9dcef392d68/depsi/classification.py#L13), which is a dedicated PSI processor. Considering SARXarray is focusing on digesting large stacks to Xarray, we consider to remove this function in the future. For now, I added a deprecation warning for this function.
