### 2026-06-25 573ca5e
Add SNAP data and metadata reader
