+++ sarxarray/__init__.py
@@ -1,5 +1,11 @@
 from sarxarray import stack
-from sarxarray._io import from_binary, from_dataset, read_metadata, to_binary
+from sarxarray._io import (
+    from_binary,
+    from_dataset,
+    from_snap_dataset,
+    read_metadata,
+    to_binary,
+)
 from sarxarray.utils import complex_coherence, crop, multi_look
 
 __all__ = (
@@ -7,6 +13,7 @@ __all__ = (
     "from_binary",
     "to_binary",
     "from_dataset",
+    "from_snap_dataset",
     "read_metadata",
     "multi_look",
     "complex_coherence",
