+++ sarxarray/_io.py
@@ -1,9 +1,10 @@
+import json
 import logging
 import math
 import os
 import re
 from collections import defaultdict
-from datetime import datetime
+from datetime import datetime, timedelta
 from pathlib import Path
 from typing import Literal
 
@@ -14,15 +15,19 @@ import xarray as xr
 
 from .conf import (
     META_ARRAY_KEYS,
+    META_ARRAY_SHAPES_SNAP,
     META_FLOAT_KEYS,
     META_INT_KEYS,
     META_UNIT_CONVERSION_MULTIPLICATION_KEYS_DORIS4,
     META_UNIT_CONVERSION_MULTIPLICATION_KEYS_DORIS5,
+    META_UNIT_CONVERSION_MULTIPLICATION_KEYS_SNAP,
     RE_PATTERNS_DORIS4,
     RE_PATTERNS_DORIS5,
     RE_PATTERNS_DORIS5_IFG,
+    RE_PATTERNS_SNAP,
     TIME_FORMAT_DORIS4,
     TIME_FORMAT_DORIS5,
+    TIME_FORMAT_SNAP,
     TIME_STAMP_KEY,
     _dtypes,
     _memsize_chunk_mb,
@@ -37,7 +42,7 @@ def from_dataset(ds: xr.Dataset) -> xr.Dataset:
     This function create tasks graph converting the two data variables of complex data:
     `real` and `imag`, to three variables: `complex`, `amplitude`, and `phase`.
 
-    The function is intented for an SLC stack in `xr.Dataset` loaded from a Zarr file.
+    The function is intended for an SLC stack in `xr.Dataset` loaded from a Zarr file.
 
     For other datasets, such as lat, lon, etc., please use `xr.open_zarr` directly.
 
@@ -170,6 +175,227 @@ def from_binary(
     if vlabel == "complex":
         ds_stack = ds_stack.slcstack._get_amplitude()
         ds_stack = ds_stack.slcstack._get_phase()
+
+    return ds_stack
+
+
+def from_snap_dataset(snap_znap_archives: list[str, Path]) -> xr.Dataset:
+    """Read an SLC stack from a list of ZNAP archives produced by SNAP.
+
+    SNAP produces .znap-archives, which are very similar to the .zarr
+    archives used by sarxarray except for a few key differences:
+    - ZNAPs use a local xy-coordinate system instead of radar coordinates
+    - ZNAPs separate the real and complex part of the observation
+    - SNAP produces one ZNAP for each image
+
+    This function takes a series of ZNAPs and converts them to the
+    standard sarxarray output: one .zarr archive with coordinates
+    azimuth and range (corrected for the offset of the local coordinate
+    system) and time, and the complex data layer for the observations.
+    Additional data layers present in the ZNAP archives with xy-coordinates
+    are passed on as well, either as (azimuth, range, time) variables (if
+    present at a daughter epochs) or as (azimuth, range) variables (if
+    present only at the mother epoch (which is automatically detected)).
+
+    ZNAP archives can also contain 'tiepoint grids' with xt/yt coordinates.
+    These layers are not passed on, as they do not fit into the azimuth/range
+    coordinate system of sarxarray.
+
+    Parameters
+    ----------
+    snap_znap_archives: list[str, Path]
+        List of .znap archives to be read into an xarray Dataset
+
+    Returns
+    -------
+    xr.Dataset
+        Dataset containing:
+            - the coordinates azimuth/range (corrected for cropping offsets) and time
+            - complex data layer based on ZNAPs i and q data layers
+            - amplitude and phase data layers as calculated from the complex data
+            - any extra data layers present in the ZNAPs at a daughter acquisition as
+                (azimuth, range, time) data variables
[…]
