### 2026-06-25 a1329d7
Fix ruff issues
