+++ sarxarray/conf.py
@@ -79,7 +79,9 @@ RE_PATTERNS_DORIS5_IFG = {
 }
 # Regular expressions for reading metadata from SNAP
 RE_PATTERNS_SNAP = {
-    "sar_processor": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.Processing_system_identifier",
+    "sar_processor": (
+        r"[\d]+.Abstracted_Metadata.attributes.[\d]+.Processing_system_identifier"
+    ),
     "product_type": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.PRODUCT_TYPE",
     "orbit": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.REL_ORBIT",
     "pass_direction": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.PASS",
@@ -87,25 +89,54 @@ RE_PATTERNS_SNAP = {
     "image_mode": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.ACQUISITION_MODE",
     "polarisation": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.mds[\d]+_tx_rx_polar",
     "range_pixel_spacing": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.range_spacing",
-    "azimuth_pixel_spacing": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.azimuth_spacing",
+    "azimuth_pixel_spacing": (
+        r"[\d]+.Abstracted_Metadata.attributes.[\d]+.azimuth_spacing"
+    ),
     "radar_frequency": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.radar_frequency",
     "sensor_platform": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.MISSION",
-    "pulse_repetition_frequency": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.pulse_repetition_frequency",
-    "first_azimuth_time": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.first_line_time",
-    "azimuth_time_interval": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.line_time_interval",
-    "total_azimuth_bandwidth": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.azimuth_bandwidth",
+    "pulse_repetition_frequency": (
+        r"[\d]+.Abstracted_Metadata.attributes.[\d]+.pulse_repetition_frequency"
+    ),
+    "first_azimuth_time": (
+        r"[\d]+.Abstracted_Metadata.attributes.[\d]+.first_line_time"
+    ),
+    "azimuth_time_interval": (
+        r"[\d]+.Abstracted_Metadata.attributes.[\d]+.line_time_interval"
+    ),
+    "total_azimuth_bandwidth": (
+        r"[\d]+.Abstracted_Metadata.attributes.[\d]+.azimuth_bandwidth"
+    ),
     "weighting_azimuth": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.azimuth_looks",
-    "first_range_time": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.slrTimeToFirstValidPixel",
-    "range_sampling_rate": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.range_sampling_rate",
-    "total_range_bandwidth": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.range_bandwidth",
-    "orbit_time": r"[\d]+.Abstracted_Metadata.elements.[\d]+.Orbit_State_Vectors.elements.[\d]+.orbit_vector[\d]+.attributes.[\d]+.time",
[…]
