### 2026-06-25 8d0d7ae
Add tests, fix name interference
