+++ sarxarray/_io.py
@@ -228,7 +228,7 @@ def from_snap_dataset(snap_znap_archives: list[str, Path]) -> xr.Dataset:
     """
     # Check if snap_znap_archives is a non empty Iterable and not a string
     if not hasattr(snap_znap_archives, "__iter__") or isinstance(
-            snap_znap_archives, str
+        snap_znap_archives, str
     ):
         raise ValueError(
             "snap_znap_archives should be a non-empty Iterable and not a string."
@@ -245,12 +245,13 @@ def from_snap_dataset(snap_znap_archives: list[str, Path]) -> xr.Dataset:
 
     for file in snap_znap_archives:
         data = xr.open_zarr(file, consolidated=False)
-        data_layers = data.original_raster_data_node_order
-        cur_epoch = data_layers[0].split("_")[-1]
-        epoch = datetime.strptime(cur_epoch, "%d%b%Y").strftime("%Y%m%d")
+        cur_epoch = data.attrs["time_coverage_start"]
+        time_stamp = datetime.strptime(cur_epoch, "%Y-%m-%dT%H:%M:%S.%fZ")
+        epoch = time_stamp.strftime("%Y%m%d")
 
         # there are two types of layers: full data layers [x, y], and tiepoint [xt, yt]
         # we only preserve the full data layers, as the others cannot be placed
+        data_layers = data.attrs["original_raster_data_node_order"]
         cleaned_data_layer = {}
         for layer in data_layers:
             dims = data[layer].dims
@@ -261,15 +262,16 @@ def from_snap_dataset(snap_znap_archives: list[str, Path]) -> xr.Dataset:
                     cleaned_name = cleaned_name.replace(f"_{pol}", "")
 
                 cleaned_data_layer[cleaned_name] = layer
-                if cleaned_name in [
-                    "latitude", "longitude", "elevation"
-                ] and mother_epoch is None:  # these are expected only at the mother
+                if (
+                    cleaned_name in ["latitude", "longitude", "elevation"]
+                    and mother_epoch is None
+                ):  # these are expected only at the mother
                     mother_epoch = epoch
 
         full_data[epoch] = {
             "data": data,
             "cleaned_data_layers": cleaned_data_layer,
-            "filename": file
+            "filename": file,
         }
 
     # sort the provided files chronologically
@@ -293,11 +295,11 @@ def from_snap_dataset(snap_znap_archives: list[str, Path]) -> xr.Dataset:
     coords = {
         "azimuth": range(
             metadata["first_line_number"],
-            metadata["first_line_number"] + metadata["number_of_lines"]
+            metadata["first_line_number"] + metadata["number_of_lines"],
         ),
         "range": range(
             metadata["first_pixel_number"],
-            metadata["first_pixel_number"] + metadata["number_of_pixels"]
+            metadata["first_pixel_number"] + metadata["number_of_pixels"],
         ),
         "time": sorted_epochs,
     }
@@ -308,8 +310,10 @@ def from_snap_dataset(snap_znap_archives: list[str, Path]) -> xr.Dataset:
     for epoch in sorted_epochs:
         cplx = (
             full_data[epoch]["data"][full_data[epoch]["cleaned_data_layers"]["i"]].data
-            + 1j *
-            full_data[epoch]["data"][full_data[epoch]["cleaned_data_layers"]["q"]].data
+            + 1j
+            * full_data[epoch]["data"][
+                full_data[epoch]["cleaned_data_layers"]["q"]
+            ].data
         )
 
         if slcs is None:
@@ -339,7 +343,7 @@ def from_snap_dataset(snap_znap_archives: list[str, Path]) -> xr.Dataset:
                             (
                                 metadata["number_of_lines"],
                                 metadata["number_of_pixels"],
-                                1
+                                1,
                             )
                         )
                     else:
[…]
