### 2026-06-26 a4cf66b
update the behavior when inferring data layer names from znap
