+++ sarxarray/conf.py
@@ -26,6 +26,7 @@ RE_PATTERNS_DORIS4 = {
     "weighting_range": r"Weighting_range:\s+(.+)",
     "first_azimuth_time": r"First_pixel_azimuth_time \(UTC\):\s+(.+)",
 }
+
 # Regular expressions for reading metadata from DORIS5 files
 RE_PATTERNS_DORIS5 = {
     "sar_processor": r"SAR_PROCESSOR:\s+(.+)",
@@ -70,11 +71,13 @@ RE_PATTERNS_DORIS5 = {
         r"Scene_centre_longitude:" r"\s+([-+]?\d+(?:\.\d+)?(?:[eE][-+]?\d+)?)"
     ),
 }
+
 # Regular expressions for reading metadata from DORIS5 interferogram files
 RE_PATTERNS_DORIS5_IFG = {
     "number_of_lines": r"Number of lines \(multilooked\):\s+(\d+)",
     "number_of_pixels": r"Number of pixels \(multilooked\):\s+(\d+)",
 }
+
 # Regular expressions for reading metadata from SNAP
 RE_PATTERNS_SNAP = {
     "sar_processor": (
@@ -139,12 +142,14 @@ RE_PATTERNS_SNAP = {
     ),
     "first_line_number": r"[\d]+.Abstracted_Metadata.attributes.[\d]+.subset_offset_y",
 }
+
 # Regular expressions for renaming SNAP Zarr file (ZNAP) datalayers.
 RE_PATTERNS_SNAP_DATALAYER = {
     "i": r"^i_(VV|VH|HH|HV)_\d{1,2}[A-Za-z]{3}\d{4}$",  # e.g., i_VV_19Mar2023
     "q": r"^q_(VV|VH|HH|HV)_\d{1,2}[A-Za-z]{3}\d{4}$",  # e.g., q_VV_19Mar2023
     "h2ph": r"^h2ph_(VV|VH|HH|HV)_\d{1,2}[A-Za-z]{3}\d{4}$",  # e.g., h2ph_VV_19Mar2023
 }
+
 # Float keys in metadata. They are used to regulate the metadata read as strings
 # Some of these are from DORIS5 only
 META_FLOAT_KEYS = [
@@ -162,6 +167,7 @@ META_FLOAT_KEYS = [
     "scene_centre_latitude",
     "scene_centre_longitude",
 ]
+
 # Integer keys in metadata. They are used to regulate the metadata read as strings
 META_INT_KEYS = [
     "deramp",  # from here DORIS5 only
@@ -172,6 +178,7 @@ META_INT_KEYS = [
     "first_pixel_number",  # from here SNAP
     "first_line_number",
 ]
+
 # Array keys in metadata and their format. Requires re.findall instead of re.match
 # Expects 2D arrays, and a callable variable type as value associated with each key
 META_ARRAY_KEYS = {
@@ -181,6 +188,7 @@ META_ARRAY_KEYS = {
     "orbit_velocity": float,
     "polarisations": str,
 }
+
 # SNAP returns flattened 1D arrays, so we need to tell it the shapes. The auto
 # dimension gets expanded to the total number of inputs, and is required
 META_ARRAY_SHAPES_SNAP = {
@@ -189,6 +197,7 @@ META_ARRAY_SHAPES_SNAP = {
     "orbit_velocity": ("auto", 3),
     "polarisations": ("auto", 1),
 }
+
 # Some keys are not read in in SI units. The following dictionary specifies those
 # keys, and the factor they should be multiplied by to restore them to SI units
[…]
