### 2026-06-26 4da10dd
change logic of handling other dimensions than x, and y
