### 2026-06-26 53b1eb7
logic for identifying mother epoch
