+++ sarxarray/_io.py
@@ -30,6 +30,7 @@ from .conf import (
     TIME_FORMAT_DORIS5,
     TIME_FORMAT_SNAP,
     TIME_STAMP_KEY,
+    ZNAP_DATA_VAR_MOTHER,
     _dtypes,
     _memsize_chunk_mb,
 )
@@ -241,173 +242,89 @@ def from_snap_dataset(snap_znap_archives: list[str, Path]) -> xr.Dataset:
 
     # in the .znaps, the dimension x is range, y is azimuth
     # xt and yt are interpolated values at coarse resolution
-    full_data = {}
-    mother_epoch = None
+    ds_stack = None
+    is_mother = False
 
+    # Loop over all ZNAP archives and read into ds_stack
     for file in snap_znap_archives:
-        # open one ZNAP archive
-        data = xr.open_zarr(file, consolidated=False)
-
-        # there are two types of layers: full data layers [x, y], and tiepoint [xt, yt]
-        # we only preserve the full data layers, as the others cannot be placed
-        dims_non_xy = [dim for dim in data.dims if dim not in ["x", "y"]]
-        data = data.drop_dims(dims_non_xy)  # Drop all non-x/y dims
+        # Read the ZNAP archive and check if it is a mother epoch
+        data, is_mother = _read_one_znap_archive(file)
 
         # Get current epoch
         cur_epoch = data.attrs["time_coverage_start"]
         time_stamp = datetime.strptime(cur_epoch, "%Y-%m-%dT%H:%M:%S.%fZ")
         epoch = time_stamp.strftime("%Y%m%d")
 
-        # Rename the data variables according to RE_PATTERNS_SNAP_DATALAYER
-        cleaned_data_layers = {}
-        for layer in data.data_vars.keys():
-            cleaned_name = layer
-            for key, pattern in RE_PATTERNS_SNAP_DATALAYER.items():
-                if re.match(pattern, layer):
-                    cleaned_name = key
-            cleaned_data_layers[cleaned_name] = layer
+        # If mother epoch
+        # keep variables ZNAP_DATA_VAR_MOTHER separately
+        # Assign an all zero h2ph variable
+        if is_mother:
+            data_mother = data[ZNAP_DATA_VAR_MOTHER]
+            data = data.drop_vars(ZNAP_DATA_VAR_MOTHER)
+            data = data.assign(
+                {"h2ph": (("y", "x"), np.zeros((data.sizes["y"], data.sizes["x"])))}
+            )
 
-        # Assign data
-        full_data[epoch] = {
-            "data": data,
-            "cleaned_data_layers": cleaned_data_layers,
-            "filename": file,
-        }
+        # Assign to ds_stack along the time dimension
+        if ds_stack is None:  # first epoch, initialize ds_stack
+            ds_stack = data.expand_dims(time=[time_stamp])
+            # Drop attrs inherited from the first epoch
+            ds_stack.attrs = {}
+            # Always initialize mother_epoch to None
+            # If first epoch is mother, it will be set to epoch below
+            ds_stack = ds_stack.assign_attrs({"mother_epoch": None})
+        else:
+            ds_stack = xr.concat(
+                [ds_stack, data.expand_dims(time=[time_stamp])],
+                dim="time",
+                combine_attrs="override",  # override the attrs of the first epoch
+            )
 
-        # Check if mother epoch
-        if any(
-            layer in cleaned_data_layers
-            for layer in ["latitude", "longitude", "elevation"]
-        ):
-            mother_epoch = epoch
+        # Modify the ds_stack if mother epoch
+        if is_mother:
+            ds_stack = ds_stack.assign_attrs({"mother_epoch": epoch})
 
[…]
