### 2026-06-26 af78740
handle stack without mother
