### 2026-06-29 b62dcb6
shortening chain calls
