### 2026-06-29 873d7c5
record epoch and files in a dict for sorting
