### 2026-06-29 c4ab050
output annotation
Co-authored-by: Copilot Autofix powered by AI <175728472+Copilot@users.noreply.github.com>
