### 2026-06-29 b851c51
remove redundant re pattern match check
