### 2026-06-29 837a5af
sort built ds_stack by time
