### 2026-07-02 4030ece
update changelog skills
