### 2026-07-02 57f77f5
add a section in CONTRIBUTING.md on how to make a release
