### 2026-07-02 cd2f308
moving agent instructions from issue template to CONTRIBUTING guide
