+++ .github/ISSUE_TEMPLATE/release-checklist.md
@@ -14,7 +14,5 @@ I will make a new release by preforming the following steps:
 - [ ] Step2: Update `docs/CHANGELOG.md` with new change information
 - [ ] Step3: Create release/tag at https://github.com/TUDelftGeodesy/sarxarray/releases
 
-Hints:
+Hint(s):
 - Step 1 and 2 can be done semi-automatically by calling agent skill, see the **Making a release** section in [CONTRIBUTING.md](../docs/CONTRIBUTING.md). However the changes will always be reviewed and commited by human.
-- The agent skill can be executed in GenAI prompt, e.g. GitHub Copilot Chat, by `/release-changelog vX.Y.Z` command, where `vX.Y.Z` is the new version label.
-- Alternatively, one can run the helper script `bash .github/skills/release-changelog/scripts/generate_release_section.sh vX.Y.Z` to get bullets for the new release section.
