### 2026-07-02 8bd6d8e
add binary writing to joss paper
