### 2026-07-02 TUDelftGeodesy/sarxarray#99 @rogerkuou
@weiji14  In addition to the mentioned change above, in PR #104 we also mentioned a new functionality of writing to binary, which is `to_binary`
