### 2026-07-03 b481c66
fix two typos
