### 2026-07-06 7556a7c
Rework timestamp to read from metadata file
