### 2026-07-06 2dcfd5d
Fix ruff issue
