### 2026-07-07 6fee3b4
fix missing references
