+++ paper/paper.bib
@@ -86,7 +86,8 @@ year = {2019}
   number={1},
   pages={6--43},
   year={2013},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={https://doi.org/10.1109/MGRS.2013.2248301}
 }
 
 @article{LUMBANGAOL2025117551,
@@ -124,3 +125,24 @@ doi={10.1109/TGRS.2025.3600893}}
   pages={10732-10735},
   keywords={Heart;Soil measurements;Atmospheric modeling;Time series analysis;Stars;Soil;Predictive models},
   doi={10.1109/IGARSS53475.2024.10642504}}
+
+  @article{hoyer2017xarray,
+  title     = {xarray: {N-D} labeled arrays and datasets in {Python}},
+  author    = {Hoyer, S. and J. Hamman},
+  journal   = {Journal of Open Research Software},
+  volume    = {5},
+  number    = {1},
+  year      = {2017},
+  publisher = {Ubiquity Press},
+  doi       = {10.5334/jors.148},
+  url       = {https://doi.org/10.5334/jors.148}
+}
+
+@article{rocklin2015,
+  author = {Rocklin, Matthew},
+  title = {Dask: Parallel Computation with Blocked algorithms and Task Scheduling},
+  journal = {SciPy 2015},
+  year = {2015},
+  doi = {10.25080/Majora-7b98e3ed-013},
+  url = {https://doi.org/10.25080/Majora-7b98e3ed-013}
+}
