+++ paper/paper.bib
@@ -87,7 +87,7 @@ year = {2019}
   pages={6--43},
   year={2013},
   publisher={IEEE},
-  doi={https://doi.org/10.1109/MGRS.2013.2248301}
+  doi={10.1109/MGRS.2013.2248301}
 }
 
 @article{LUMBANGAOL2025117551,
