### 2026-07-07 52cfa12
switch to shield io for a stable rendering of zenodo doi badge
