### 2026-07-08 460cf95
Re-add wildcard name stripping
