### 2026-07-08 3afa66f
Remove leftover debugger
