### 2026-07-08 2b5d353
Document other reason for ValueError
