### 2026-07-09 ef63e91
Revert "apply joss review suggestions on paper"
This reverts commit 4440818f7a9527f4ca6a7ff69bfde9ceee2696c6.
