+++ paper/paper.bib
@@ -146,3 +146,11 @@ doi={10.1109/TGRS.2025.3600893}}
   doi = {10.25080/Majora-7b98e3ed-013},
   url = {https://doi.org/10.25080/Majora-7b98e3ed-013}
 }
+
+@misc{xarray-sentinel,
+  author = {A. Amici and et al.},
+  title = {xarray-sentinel},
+  year = {2024},
+  url = {https://github.com/bopen/xarray-sentinel},
+  note = {GitHub repository, accessed 2026-07-09}
+}
