+++ paper/paper.bib
@@ -148,7 +148,7 @@ doi={10.1109/TGRS.2025.3600893}}
 }
 
 @misc{xarray-sentinel,
-  author = {A. Amici and et al.},
+  author = {Amici, A.},
   title = {xarray-sentinel},
   year = {2024},
   url = {https://github.com/bopen/xarray-sentinel},
