+++ paper/paper.bib
@@ -148,7 +148,7 @@ doi={10.1109/TGRS.2025.3600893}}
 }
 
 @misc{xarray-sentinel,
-  author = {Amici, A.},
+  author = {Amici, A and others},
   title = {xarray-sentinel},
   year = {2024},
   url = {https://github.com/bopen/xarray-sentinel},
