### 2026-07-10 0776fd8
update changelog and version number in pyproject.toml
