+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "sarxarray"
-version = "1.2.3"
+version = "1.3.0"
 requires-python = ">=3.10"
 dependencies = [
     "dask[complete]",
