### 2026-07-13 ffae6ce
reduce function complexity
