+++ sarxarray/_io.py
@@ -181,7 +181,7 @@ def from_binary(
     return ds_stack
 
 
-def from_snap_dataset(snap_znap_archives: list[str, Path]) -> xr.Dataset:
+def from_snap_dataset(snap_znap_archives: list[str | Path]) -> xr.Dataset:
     """Read an SLC stack from a list of ZNAP archives produced by SNAP.
 
     SNAP produces .znap-archives, which are very similar to the .zarr
@@ -208,7 +208,7 @@ def from_snap_dataset(snap_znap_archives: list[str, Path]) -> xr.Dataset:
 
     Parameters
     ----------
-    snap_znap_archives: list[str, Path]
+    snap_znap_archives: list[str | Path]
         List of .znap archives to be read into an xarray Dataset
 
     Returns
@@ -230,17 +230,8 @@ def from_snap_dataset(snap_znap_archives: list[str, Path]) -> xr.Dataset:
         If the timestamp of the image cannot be found in the metadata
         If multiple mother images are identified
     """
-    # Check if snap_znap_archives is a non empty Iterable and not a string
-    if not hasattr(snap_znap_archives, "__iter__") or isinstance(
-        snap_znap_archives, str
-    ):
-        raise ValueError(
-            "snap_znap_archives should be a non-empty Iterable and not a string."
-            "If you have only one file, please put it in a list, e.g. "
-            "snap_znap_archives=[file]."
-        )
-    if len(snap_znap_archives) == 0:
-        raise ValueError("snap_znap_archives should be a non-empty Iterable.")
+    # Validate input for from_snap_dataset
+    _validate_snap_znap_archives(snap_znap_archives)
 
     # Loop over all ZNAP archives and read into ds_stack
     ds_stack = None  # Stack of all epochs as xr.Dataset
@@ -254,29 +245,20 @@ def from_snap_dataset(snap_znap_archives: list[str, Path]) -> xr.Dataset:
         data, is_mother = _read_one_znap_archive(file)
 
         # Get current epoch
+        # Mother and daughter epochs need to be treated differently
         metadata = read_metadata(f"{file}/SNAP/product_metadata.json", driver="snap")
-        if is_mother:
-            cur_epoch_raw = metadata["mother_file"]
-        else:
-            cur_epoch_raw = metadata["daughter_file"][0][0]  # pick the first one
-        cur_epochs = cur_epoch_raw.split("_")
-        # Match on YYYYMMDDTHHMMSS, 8 digits, letter T, 6 digits
-        matches = list(filter(re.compile(r"^[\d]{8}T[\d]{6}$").match, cur_epochs))
-        if len(matches) == 0:
-            raise ValueError(
-                f"Could not find epoch timestamp YYYYMMDDTHHMMSS in {cur_epoch_raw}."
-            )
-        time_stamp = datetime.strptime(matches[0], "%Y%m%dT%H%M%S")
-        epoch = time_stamp.strftime("%Y%m%d")
+        time_stamp, epoch = _extract_snap_epoch(metadata, is_mother)
 
         # register the epoch and file in a dictionary for later use
         epoch_file_dict[epoch] = file
 
-        # If mother epoch
-        # keep variables ZNAP_DATA_VAR_MOTHER separately
-        # Assign an all zero h2ph variable
-        if is_mother:
-            # check if mother epoch has already been found
+        # Append all daughter epochs to ds_stack
+        # Mother epoch will be added later
+        if not is_mother:
+            ds_stack = _append_one_daughter(ds_stack, data, time_stamp)
+        else:
+            # If mother epoch, keep variables ZNAP_DATA_VAR_MOTHER separately
+            # Assign an all zero h2ph variable
             if data_mother is not None:
                 raise ValueError(
[…]
