### 2026-07-13 ccebf3f
fix sonarcloud warnings
