### 2026-07-13 c188fd4
Update paper/paper.md
Co-authored-by: FreekvanLeijen <f.j.vanleijen@tudelft.nl>
