### 2026-07-13 a08e953
Update paper/paper.md
Co-authored-by: SarahAlidoost <55081872+SarahAlidoost@users.noreply.github.com>
