### 2026-07-13 ce47f92
refactor _flatten_snap_json_metadata
