### 2026-07-13 da7e9d6
changelog for the new release
