### 2026-07-15 48496c8
Handle and test case with only mother ZNAP
