+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "sarxarray"
-version = "1.3.1"
+version = "1.4.0b1"
 requires-python = ">=3.10"
 dependencies = [
     "dask[complete]",
