### 2026-08-20 TUDelftGeodesy/sarxarray#121 @rogerkuou
JOSS review comments on the paper
https://github.com/openjournals/joss-reviews/issues/10492#issuecomment-5332270462

- [x] Page 1: “e.g. civil-infrastructure stability” needs a comma after “e.g.”: “e.g., civil-infrastructure stability.”
- [x] Page 1: “a Xarray extension” should be “an Xarray extension.”
- [x] Page 1: “two well-established Python libraries Xarray ... and Dask” needs punctuation after “libraries,” e.g. “two well-established Python libraries, Xarray ... and Dask.”
- [x] Page 1: “xarray-sentinel(Amici & others, 2024)” is missing a space before the citation.
- [x] Page 2: “This design is motivated by Xarray community’s recommendation” is missing an article: “the Xarray community’s recommendation.”
- [x] Page 3: “in Pull Requests reviews” should be “in Pull Request reviews” or “in reviews of Pull Requests.”
- [x] Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "dutch" and others.
