### 2026-08-20 #10492 @rogerkuou
> [@rogerkuou](https://github.com/rogerkuou)
>
> * Page 1: “e.g. civil-infrastructure stability” needs a comma after “e.g.”: “e.g., civil-infrastructure stability.”
> * Page 1: “a Xarray extension” should be “an Xarray extension.”
> * Page 1: “two well-established Python libraries Xarray ... and Dask” needs punctuation after “libraries,” e.g. “two well-established Python libraries, Xarray ... and Dask.”
> * Page 1: “xarray-sentinel(Amici & others, 2024)” is missing a space before the citation.
> * Page 2: “This design is motivated by Xarray community’s recommendation” is missing an article: “the Xarray community’s recommendation.”
> * Page 3: “in Pull Requests reviews” should be “in Pull Request reviews” or “in reviews of Pull Requests.”
> * Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "dutch" and others.

Thanks @kthyng ! I have applied all your suggestions via PR https://github.com/TUDelftGeodesy/sarxarray/pull/122 .

Please let me know if there are more actions needed from my side.
